small corrections

+++ include/maze_generator_recursive_backtracking.h
@@ -13,11 +13,11 @@ struct node;
 typedef struct node node;
 
 #ifdef BIG_MAZE
-const u64 get_next(node* maze, u64 node_idx, u32 height, u32 width);
+const u64 get_next(node* maze, u64 node_idx, u64 height, u64 width);
 #else
 const u32 get_next(node* maze, u32 node_idx, u32 height, u32 width);
 #endif
 
-char* generate_maze(u32 height, u32 width);
+char* generate_maze(u64 height, u64 width);
 
 #endif //DOYOUWANTTOCMYALGORITHMS__MAZE_GENERATOR_RECURSIVE_BACKTRACKING_H
